backend/requirements.txt
Flask==3.0.0
firebase-admin>=6.0.0
sentence-transformers>=2.2.0
google-generativeai>=0.3.0
python-dotenv==1.0.0
flask-cors==4.0.0
numpy>=1.24.0
